feat: AVFoundation encoder uses EncodingController
